+++ examples/nextjs-comments-overlay/src/components/comments/PinnedComposer.tsx
@@ -1,9 +1,10 @@
 "use client";
 
 import styles from "./Pinned.module.css";
-import { PointerEventHandler } from "react";
+import { PointerEventHandler, useRef } from "react";
 
 import { Composer, ComposerProps } from "@liveblocks/react-ui";
+import { useNearEdge } from "@/lib/useNearEdge";
 
 type Props = {
   user: Liveblocks["UserMeta"]["info"];
@@ -21,17 +22,32 @@ export function PinnedComposer({
   onComposerSubmit,
   ...props
 }: Props) {
+  // Flip pinnedContent away from edge of screen
+  const ref = useRef(null);
+  const { nearRightEdge, nearBottomEdge } = useNearEdge(ref);
+
   return (
-    <div className={styles.pinned} {...props}>
+    <div ref={ref} className={styles.pinned} {...props}>
       <div
         className={styles.avatarPin}
         onPointerDown={onPointerDown}
         onPointerMove={onPointerMove}
         onPointerUp={onPointerUp}
+        data-draggable={true}
       >
-        <img src={user.avatar} alt={user.name} width="28px" height="28px" />
+        <img
+          src={user.avatar}
+          alt={user.name}
+          width="28px"
+          height="28px"
+          draggable={false}
+        />
       </div>
-      <div className={styles.pinnedContent}>
+      <div
+        className={styles.pinnedContent}
+        data-flip-vertical={nearBottomEdge || undefined}
+        data-flip-horizontal={nearRightEdge || undefined}
+      >
         <Composer
           onComposerSubmit={onComposerSubmit}
           onClick={(e) => {