+++ packages/liveblocks-node/src/__tests__/client.test.ts
@@ -885,6 +885,132 @@ describe("client", () => {
     }
   });
 
+  test("should return the user's inbox notifications when getInboxNotifications receives a successful response", () => {
+    const userId = "user1";
+
+    const notifications = [
+      {
+        id: "notification1",
+        kind: "thread",
+        notifiedAt: new Date().toISOString(),
+        readAt: null,
+        threadId: "thread1",
+      },
+    ];
+
+    server.use(
+      http.get(
+        `${DEFAULT_BASE_URL}/v2/users/:userId/inbox-notifications`,
+        () => {
+          return HttpResponse.json({ data: notifications }, { status: 200 });
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    return expect(client.getInboxNotifications({ userId })).resolves.toEqual({
+      data: notifications.map((notification) => ({
+        ...notification,
+        notifiedAt: new Date(notification.notifiedAt),
+        readAt: notification.readAt ? new Date(notification.readAt) : null,
+      })),
+    });
+  });
+
+  test("getInboxNotifications works with a query", async () => {
+    const userId = "user1";
+    const notifications = [
+      {
+        id: "notification1",
+        kind: "thread",
+        notifiedAt: new Date().toISOString(),
+        readAt: null,
+        threadId: "thread1",
+      },
+    ];
+
+    server.use(
+      http.get(
+        `${DEFAULT_BASE_URL}/v2/users/:userId/inbox-notifications`,
+        (res) => {
+          const url = new URL(res.request.url);
+
+          expect(url.searchParams.size).toEqual(1);
+          expect(url.searchParams.get("query")).toEqual("unread:true");
+
+          return HttpResponse.json({ data: notifications }, { status: 200 });
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    await expect(
+      client.getInboxNotifications({
+        userId,
+        query: {
+          unread: true,
+        },
+      })
+    ).resolves.toEqual({
+      data: notifications.map((notification) => ({
+        ...notification,
+        notifiedAt: new Date(notification.notifiedAt),
+        readAt: notification.readAt ? new Date(notification.readAt) : null,
+      })),
+    });
+
+    // with a string query
+    await expect(
+      client.getInboxNotifications({
+        userId,
+        query: "unread:true",
+      })
+    ).resolves.toEqual({
+      data: notifications.map((notification) => ({
+        ...notification,
+        notifiedAt: new Date(notification.notifiedAt),
+        readAt: notification.readAt ? new Date(notification.readAt) : null,
+      })),
+    });
+  });
+
+  test("should throw a LiveblocksError when getInboxNotifications receives an error response", async () => {
+    const userId = "user1";
+
+    const error = {
+      error: "USER_NOT_FOUND",
+      message: "User not found",
+    };
+
+    server.use(
+      http.get(
+        `${DEFAULT_BASE_URL}/v2/users/:userId/inbox-notifications`,
+        () => {
+          return HttpResponse.json(error, { status: 404 });
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    // This should throw a LiveblocksError
+    try {
+      // Attempt to get, which should fail and throw an error.
+      await client.getInboxNotifications({ userId });
+      // If it doesn't throw, fail the test.
+      expect(true).toBe(false);
+    } catch (err) {
+      expect(err instanceof LiveblocksError).toBe(true);
+      if (err instanceof LiveblocksError) {
+        expect(err.status).toBe(404);
+        expect(err.message).toBe(JSON.stringify(error));
+        expect(err.name).toBe("LiveblocksError");
+      }
+    }
+  });
+
   test("should get user's room notification settings", async () => {
     const userId = "user1";
     const roomId = "room1";