+++ packages/liveblocks-react/src/index.ts
@@ -47,6 +47,7 @@ export {
   useErrorListener,
   useEventListener,
   useHistory,
+  useIsInsideRoom,
   useLostConnectionListener,
   useMarkThreadAsRead,
   useMutation,
@@ -83,6 +84,7 @@ export {
 } from "./room";
 export {
   useInboxNotifications,
+  useUserThreads_experimental as useUserThreads_experimental,
   useRoomInfo,
   useUnreadInboxNotificationsCount,
   useUser,