+++ CHANGELOG.md
@@ -1,10 +1,34 @@
-## 2.4.1 (Not yet published)
+## 2.6.0
+
+### `@liveblocks/node`
+
+- Add `getInboxNotifications` method which supports an `unread` query parameter.
+
+## 2.5.1
+
+### `@liveblocks/yjs`
+
+- Fix `LiveblocksProvider` `update`/`change` event not returning `removed`
+  users.
+
+## 2.5.0
+
+### `@liveblocks/react`
+
+- Add
+  [`useIsInsideRoom`](https://liveblocks.io/docs/api-reference/liveblocks-react#useIsInsideRoom)
+  hook, useful for rendering different components inside and outside of
+  [`RoomProvider`](https://liveblocks.io/docs/api-reference/liveblocks-react#RoomProvider).
 
 ### `@liveblocks/react-lexical`
 
-- Fix a bug in `useEditorStatus` which prevented the hook from returning correct
+- Fix a bug in `useEditorStatus` which prevented it from returning a correct
   status when `LexicalPlugin` was rendered conditionally.
-- Fix remote cursors not displaying user name.
+- Fix remote cursors not displaying user names.
+
+### `@liveblocks/react-ui`
+
+- Improve event propagation in `Composer`.
 
 ## v2.4.0